+++ source/Contracts/Charge/ChargeSearchCriteriaV1Dto.cs
@@ -18,7 +18,7 @@ using System.Collections.Generic;
 // ReSharper disable once CheckNamespace - Type is shared so namespace is not determined by project structure/namespace
 namespace Energinet.Charges.Contracts.Charge
 {
-    public record SearchCriteriaV1Dto(
+    public record ChargeSearchCriteriaV1Dto(
         string ChargeIdOrName,
         List<Guid> OwnerIds,
         List<ChargeType> ChargeTypes);