+++ source/Energinet.Charges.Libraries/source/Energinet.DataHub.Charges.Clients/documents/release-notes/release-notes.md
@@ -1,5 +1,11 @@
 # Energinet.DataHub.Charges.Clients Release notes
 
+## Version 4.2.0
+
+Added `SearchChargePointsAsync` to Charges client.
+Renamed `SearchCriteriaV1Dto` to `ChargeSearchCriteriaV1Dto`.
+Throws exception instead of null if status code is not success.
+
 ## Version 4.1.1
 
 Added VatClassification and ChargeDescription to ChargesV1Dto